refactor: Notification 관련 로직 클래스로 분리
